test db is created and destroyed by ServiceTestCase
