--- app/service_tests/ServiceTestCase.py
@@ -1,9 +1,0 @@
-import unittest
-
-
-class ServiceTestCase(unittest.TestCase):
-    def setUp(self):
-        super(ServiceTestCase, self).setUp()
-
-    def tearDown(self):
-        super(ServiceTestCase, self).tearDown()
